Fixed the annealing tests with the new random number generation.
I thought I had done these already.  I'm not sure what I had done before.

Also backported some other improvements from the python3 branch, just
to make comparisons easier.

+++ TEST/regression.py
@@ -250,7 +250,10 @@ def run(homedir):
     finally:
         if ok:
             print >> sys.stderr, "All tests completed successfully!"
-            os.rmdir(tmpdir)
+            if not debug:
+                os.rmdir(tmpdir)
+            else:
+                print >> sys.stderr, "Temp dir", tmpdir, "was not removed."
         else:
             print >> sys.stderr, "Test failed. Temp dir", tmpdir, "was not removed."
